Use Clearlinux standard bootloader
Update the hypervisor to use the standard Clearlinux bootloader.
This replaces the previous mechanism where the user had to make
a copy of the bootloader and call it bootloaderx64_origin.efi.
This is no longer needed and will simplify the instructions for
how to install the ACRN hypervisor with a Clearlinux-based
Service OS

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>