devops: Add CODEOWNERS for GitHub review
Seed an initial CODEOWNERS file.  Update as owners for specific areas
are defined.

You can use a CODEOWNERS file to define individuals or teams that are
responsible for code in a repository. The listed names must have write
permissions for the repository.

Code owners are automatically requested for review when someone opens a
pull request that modifies code that they own. When someone with admin
or owner permissions has enabled required reviews, they also can
optionally require approval from a code owner before the author can
merge a pull request in the repository.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>