HV: automatically generate acrn (hypervisor) version from git
This patch changes how the acrn hypervisor version is set.
Previously, a "version.h" was constructed that included a number
of macros that were used to build up the version number. With
this patch, the version is directly taken from git (using
'git describe').

The most recent tag is used as the baseline for the version and
more information is automatically added (such as number of commits
on top of that tag, whether the working tree is dirty, etc.). For
more information on the exact format, see 'man git describe'

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>