Getting Started Guide for the UP2 board
Fix formatting and adhere to the recommendations published by
the documentation owner when writing reST documents.

Simplify the instructions by referencing the main Getting Started
Guide. This is now possible with the latest code since what
required us to modify the source code and build the components
are now parameters we can set when installing ACRN. This
simplifies the instructions *a lot*.

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>