doc: reorganize doc tree
With the addition of more HLD documents, the developer-guides index was
getting too busy, so push the HLD documents down a level.

Also, the supported hardware document is buried in the getting started
section and hard to find, so promote it.

Trusty isn't really supported (yet) so drop it from the TOC for now.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>