Hypervisor: fix new default configuration mechanism
There is a mechanism foreseen that allows one to create a new
default hypervisor configuration and re-use it very simply
with `make PLATFORM=new-platform`. The mechanism does not work
currently as it uses the new platform name to find some source
files whereas it should use either `uefi` or `sbl`.

Tracked-On: #1484

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>