hv: fix an issue in vlapic_trigger_lvt
In commit e24899d9c993 ("fix "Recursion in procedure calls found""), it
added a check if vector < 16U in vlapic_trigger_lvt() by a mistake, since
the vector variable in vlapic_trigger_lvt() is not the interrupt vector.
V2: added back the check of vector in vlapic_fire_lvt()

This patch fixed the issue mentioned above.

Tracked-On:#2598
Signed-off-by: Min He <min.he@intel.com>
Acked-by: Eddie Dong <eddie.dong@intel.com>