Doc: add space at the end of efibootmgr bootloader
This is a workaround for the issue in efi-stub: If the EFI option is loaded to a memory range which is not all zero filled,
it's possible the garbage string appended to bootloader name string.
Which could make system can't boot.

Add an extra space to the EFI option to make sure space could be used to detect the end of bootloader name string.

Once the fixing patch is merged, we don't need this extra space.
But to make old ACRN work, we'd like to keep the space here.

Signed-off-by: fuzhongl <fuzhong.liu@intel.com>
Reviewed-by: Yin Fengwei <fengwei.yin@intel.com>