doc: Known Limitation regarding PCI devices and resources assignment
ACRN assumes certain state for the PCI devices w.r.t. resources assigned to them
before the platform BIOS hands control over to itself.
When the same BIOS is used with native OS boot, there should not be a need for
native kernel to re-program device BARs or bridge windows for bridge devices.

Signed-off-by: Sainath Grandhi <sainath.grandhi@intel.com>