acrn-config: expose GPIO Chassis Interrupt to the Safety VM as INTx
This patch is to expose GPIO chassis interrupts as INTx to the safety VM for
EHL. User can configure this setting in the VM settings using the following
format:
<pt_intx desc="pt intx mapping.">
  	(phys_pin0, virt_pin0), (phys_pin1, virt_pin1), (phys_pinN, virt_pinN)
 </pt_intx>

The physical and virtual interrupt pin in each pair are separated by a
comma and enclosed in parentheses. If an integer begins with 0x or 0X,
it is hexadecimal, otherwise, it is assumed to be decimal. Example:
  <pt_intx desc="pt intx mapping.">
  	(1, 0), (0x3, 1), (0x4, 2), (5, 6) (89, 0x12)
  </pt_intx>

Tracked-On: #5220
Signed-off-by: dongshen <dongsheng.x.zhang@intel.com>