doc: add sphinx-generated index to leftnav
Sphinx generates a full-site index (headings, terms, function names,
etc.) but we haven't been linking to it.  This is a tricky way to get
the site index added to the left nav that appears to work, by creating a
genindex.rst that we can link to in the site toctree so it shows up in
the leftnav, but the generated genindex.html for the dummy genindex.rst
will be overwritten by the sphinx created genindex.html with the
full-site index.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>