hv: hypercall: hot fix call assign/deassign mmiodev could crash hv
If we assign the mmio device address and size maliciously, this may crash
the HV, SOS even the UOS. This patch is a hot fix. The formal one needs
the config tool to configure the mmio device which we could pass through
and the HV check the mmio device by it.

Tracked-On: #5555
Signed-off-by: Li Fei1 <fei1.li@intel.com>