doc: update Kconfig options documentation
Kconfig options documentation used a linking capability (:option:) but
did it in an unnamed domain that would complicate use of this capability
for other domains (such as the configuration tool options).

Create the Kconfig options in a new Kconfig domain (done in the genrest.py
script) and update references to specify this domain in the
documentation.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>