doc: update doc build instructions
We've validated doc build tool versions, so let's make sure those are
the versions the instructions say to install.  The version of doxygen
you get when you use ``sudo apt install doxygen`` may get a newer
version that may still work so let's tell them that.

Also, we no longer use kconfig files in the document build process, so
remove mentioning that in the build documentation.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>