doc: tweak doc config for v2.3 for CI tool mismatch
All releases of ACRN use the same CI docker image for testing the
documentation build. This creates a problem because some features used
in older ACRN releases are deprecated or behave differently with newer
versions of Sphinx and other tools.  Add a check for the Sphinx version
to account for this, and update the .known-issues patterns for duplicate
declarations.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>