HV: vm_load: build vrsdp after VM image loaded
The vRSDP region of pre-launched VM is located at 0xf2400, it could be
overlapped with the memory space of the raw image so be overriden by
raw image loading. In this case VM would be failed to find ACPI table.
This patch fix this issue by loading raw image first and then building
vRSDP later.

Tracked-On: #6305

Signed-off-by: Victor Sun <victor.sun@intel.com>