dm: use "igd" parameter name when pass-thru IGD
When pass-thru IGD(Intel Graphics Device),
need to use the extra paramater "gpu" to
create the dedicated "igd-lpc" on 00:1f.0.
While when pass-thru other types of GPU, don't need this "gpu" parameter.

The term of "gpu" is too common that might mislead end-users
that it is a "must have" parameter to pass though any graphics card.
Actually the parameter is used for Intel graphics only(bdf@00:02.0).
We met this problem when we try to pass through
a third party graphics card (bdf@00:05.0),
people don't know whether need to add "gpu" parameter.

Use "igd" as the paramater name instead of "gpu".

Tracked-On: #6357

Signed-off-by: Liu,Junming <junming.liu@intel.com>