config_tool: use "igd" parameter name when pass-thru IGD
The term of "gpu" is too common that might mislead end-users
that it is a "must have" parameter to pass though any graphics card.
Actually the parameter is used for Intel graphics only(bdf@00:02.0).
We met this problem when we try to pass through
a third party graphics card (bdf@00:05.0),
people don't know whether need to add "gpu" parameter.

Use "igd" as the paramater name instead of "gpu".

Tracked-On: #6357

Signed-off-by: Liu,Junming <junming.liu@intel.com>