config_tools: schema: revert to old schema
Commit e678dc8 ("doc: pick doc updates for v2.6 release") picked the latest
scenario XML schema from the main branch, which includes both documentation
changes and new configuration entries. These new entries cause build
failures of the release branch because those new entries do not exist in
the scenario XMLs of v2.6 and the default value populator is yet to be
added for v2.7.

This patch removes those new entries to unblock the build.

Tracked-On: #5692
Signed-off-by: Junjie Mao <junjie.mao@intel.com>