hv: hide thermal interface from guests
Thermal events are delivered through lapic thermal LVT. Currently
ACRN does not support delivering those interrupts to guests by
virtual lapic. They need to be virtualized to provide guests some
thermal management abilities. Currently we just hide those thermal
program interfaces from guests, including:

1. Thermal CPUIDs and MSRs:
   CPUID.01H:EDX[22] MSR_IA32_CLOCK_MODULATION,
     MSR_IA32_THERM_INTERRUPT, MSR_IA32_THERM_STATUS,
   CPUID.06H:EAX[6] IA32_PACKAGE_THERM_INTERRUPT,
    MSR_IA32_PACKAGE_THERM_STATUS,
   CPUID.01H:ECX[8], CPUID.01H:EDX[29], CPUID.06H:EAX[0,4]

2. Thermal LVT:
    There is no way to hide thermal LVT from guests. But we need do
    something to make sure no interrupt can be actually trigered:
      - skip thermal LVT in vlapic_trigger_lvt()
      - trap-and-emulate thermal LVT in lapic-pt mode

As We have plan to introduce virtualization of thermal monitor in the
future, we use a vm flag GUEST_FLAG_VTM which is default 0 to control
the access to it. So that it can help enabling VTM in the future.

Tracked-On: #8414
Signed-off-by: Wu Zhou <wu.zhou@intel.com>
Reviewed-by: Junjie Mao <junjie.mao@intel.com>