doc: add module design guidelines
ACRN hypervisor is decomposed into a series of components,
some of which are further decomposed into several modules.
The goal of ACRN module design is to further specify the
files implementing a module, the macros and declarations
in each file and the internals of each declaration.

This patch adds the module deisgn guidelines.

Signed-off-by: Gao, Shiqing <shiqing.gao@intel.com>