doc: Document requirements for pre-launched VM
For a pre-launched VM, it is essential to include 'reboot=acpi' in
the 'bootargs' of the scenario.xml file. Additionally, if the VM
is named VM0, the grub.cfg file must contain
'module2 /boot/ACPI_VM0.bin ACPI_VM0'.

Currently, several scenario files located in misc/config_tools/data
already support 'reboot=acpi'. Furthermore, the postinst Debian
script automatically adds the 'module2' command to grub.cfg if the
scenario is 'partitioned'. Therefore, by using the provided
scenarios, the pre-launched VM can be used without any additional
configuration.

However, when using a custom scenario file, there is a risk of
omitting the necessary parameter and command for a pre-launched VM.
To avoid such errors, it is useful to document these requirements.

Signed-off-by: Daichi Fukui <daichi1.fukui@toshiba.co.jp>