[FIXME] hv: Makefile: temporarily confine common files to x86 only
The common files for now contains x86 specific code and needs
refactoring. To increase parallelism, we temporarily confine the common
files to x86 architecture while we tried to enable riscv architecture.

As we progress, this workaround should be completely reverted in future
milestones when we moved all common files to actual common scope.

Tracked-On: #8782
Signed-off-by: Yifan Liu <yifan1.liu@intel.com>
Acked-by: Wang, Yu1 <yu1.wang@intel.com>