dump: dump hv image delta for risc-v
dump hypervisor image load address delta for risc-v.

Note: using the riscv64-linux-gnu-addr2line to parse the `Host Call Trace` may
not accurate when the complier using `O2` to build the hypervisor. In this case,
it'd better to use riscv64-linux-gnu-objdump to `disassemble -D acrn.out` to
parse the `Host Call Trace` manually to change the complier to use `O1` to build
the hypervisor.

Tracked-On: #8805
Signed-off-by: Fei Li <fei1.li@intel.com>
Reviewed-by: Jian Jun Chen <jian.jun.chen@intel.com>